web/requirements.txt
asgiref==3.8.1
Django==4.2.11
psycopg2-binary==2.9.10
requests==2.31.0
djangorestframework==3.14.0
openai==1.64.0
sqlparse==0.5.3
tzdata==2025.1
python-dotenv==1.0.1
